refactor(luck_insight): align structure with architecture
