+++ apps/luck_insight/lib/features/fortune/presentation/luck_theme.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class LuckColors {
+  static const background = Color(0xFFFFF7EE);
+  static const card = Color(0xFFFFFFFF);
+  static const accent = Color(0xFFFF8C32);
+  static const textPrimary = Color(0xFF3C2D25);
+  static const textSecondary = Color.fromARGB(153, 60, 45, 37);
+  static const accentBorder = Color.fromARGB(38, 255, 140, 50);
+  static const accentFill = Color.fromARGB(41, 255, 140, 50);
+  static const softShadow = Color.fromARGB(10, 0, 0, 0);
+}
